migration-pipeline/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
Pillow==10.3.0
langchain==0.2.1
langchain-community==0.2.1
